+++ src/main/java/org/realityforge/guiceyloops/server/AbstractServerTest.java
@@ -139,6 +139,11 @@ public abstract class AbstractServerTest
 
   protected void shutdownTransactionSynchronizationRegistry()
   {
+    resetJndiContext();
+  }
+
+  protected void resetJndiContext()
+  {
     try
     {
       TestInitialContextFactory.reset();